add periodic job for updating gists

+++ src/service/gist.ts
@@ -20,14 +20,16 @@ export default class GistService {
   public async UpdateGists(): Promise<void> {
     const users = await this.userRepository.ListUsers();
 
-    this.logger.info(`Updating recent gists for users ${users}`);
+    this.logger.info(`Updating recent gists for users [${users}]`);
 
     for (const user of users) {
       const previous = await this.gistRepository.GetRecent(user);
 
-      await this.gistRepository.SaveRecents(user, previous);
+      if (previous !== []) {
+        await this.gistRepository.SaveRecents(user, previous);
 
-      await this.gistRepository.DeleteRecent(user);
+        await this.gistRepository.DeleteRecent(user);
+      }
 
       const gists = await this.githubService.RequestGists(user);
 
@@ -35,8 +37,8 @@ export default class GistService {
         `Got ${gists.length} gists from github for user ${user}`
       );
 
-      const newGists = gists.filter((gist) => {
-        const exists = this.gistRepository.Exists(user, gist);
+      const newGists = await this.filterAsync(gists, async (gist) => {
+        const exists = await this.gistRepository.Exists(user, gist);
         return !exists;
       });
 
@@ -45,12 +47,20 @@ export default class GistService {
       await this.gistRepository.AddRecent(user, newGists);
     }
 
-    this.logger.info(`Done updating for users ${users}`);
+    this.logger.info(`Done updating for users [${users}]`);
   }
 
   public async GetRecent(user: string): Promise<GistResponse[]> {
     const recent = await this.gistRepository.GetRecent(user);
-
     return recent;
+  }
+
+  private mapAsync(array, callbackfn) {
+    return Promise.all(array.map(callbackfn));
+  }
+
+  private async filterAsync(array, callbackfn) {
+    const filterMap = await this.mapAsync(array, callbackfn);
+    return array.filter((value, index) => filterMap[index]);
   }
 }
